requirements.txt
django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.12
requests==2.33.1
python-dotenv==1.2.2
pytest==9.0.3
pytest-django==4.12.0
